test(slack-bot): add hello text test
